requirements.txt
ollama
PyMuPDF
python-docx
python-pptx
Pillow
PyYAML
